Move PublicUser, PublicUserSignInAttempt, PIN/token helpers to models/public_user.py
